Moved factories into module
