+++ config/build.rs
@@ -11,13 +11,13 @@
 
 use crate::executable;
 use crate::permission;
-use crate::executable::AutoExecutableFactory;
-use crate::permission::AutoPermissionFactory;
+use crate::executable::factory::AutoExecutableFactory;
+use crate::permission::factory::AutoPermissionFactory;
 
 /// The method to use to find the [Executable][eb] to run
 ///
 /// [eb]: executable::Executable
-pub const EXECUTABLE_FACTORY: AutoExecutableFactory = executable::from_commandline;
+pub const EXECUTABLE_FACTORY: AutoExecutableFactory = executable::factory::from_commandline;
 
 /// The method to use to find the current [Permission][p]s the user has
 ///
@@ -27,7 +27,7 @@ pub const EXECUTABLE_FACTORY: AutoExecutableFactory = executable::from_commandli
 /// Primary GID, and Secondary Groups.
 ///
 /// [p]: permission::Permission
-pub const SOURCE_PERMISSION_FACTORY: AutoPermissionFactory = permission::from_environment;
+pub const SOURCE_PERMISSION_FACTORY: AutoPermissionFactory = permission::factory::from_environment;
 /// The method to use to find the [Permission][p]s to run the executable as
 ///
 /// This is the target [Permission][p]s of the [Executable][eb]. As such, it is
@@ -36,17 +36,17 @@ pub const SOURCE_PERMISSION_FACTORY: AutoPermissionFactory = permission::from_en
 ///
 /// [p]: permission::Permission
 /// [eb]: executable::Executable
-pub const TARGET_PERMISSION_FACTORY: AutoPermissionFactory = permission::from_commandline;
+pub const TARGET_PERMISSION_FACTORY: AutoPermissionFactory = permission::factory::from_commandline;
 
 /// What command line argument number to look for for the path of the binary to
 /// execute
 ///
-/// Used by [executable::from_commandline]
+/// Used by [executable::factory::from_commandline]
 pub const EXECUTABLE_COMMANDLINE_PATH_IDX: usize = 4;
 /// What command line argument number to use as the first parameter to the
 /// program, with subsequent arguments being used in order
 ///
-/// Used by [executable::from_commandline]
+/// Used by [executable::factory::from_commandline]
 pub const EXECUTABLE_COMMANDLINE_ARG_START_IDX: usize = 5;
 
 /// What command line argument number to look at for the UID
