+++ src/executable/mod.rs
@@ -1,4 +1,4 @@
-//! Module representing executable objects
+//! Module representing [Executable] objects
 //!
 //! To run an executable with privileges, we need to know what executable to run
 //! and with what arguments. This module serves to collect and package that
@@ -9,13 +9,10 @@
 //! supply.
 //!
 //! Additionally, the module has methods for getting the [Executable] from the
-//! user. It has various functions to get it from command line arguments, from
-//! iterables, or from a presupplied structure.
+//! user. It has various functions to get it from command line arguments or from
+//! iterables.
 
-pub mod commandline;
-pub mod iterator;
-pub use commandline::from_commandline;
-pub use iterator::from_iterator;
+pub mod factory;
 
 use std::ffi::CString;
 use std::path::PathBuf;
@@ -30,45 +27,4 @@ pub struct Executable {
     path: PathBuf,
     /// The command line arguments to pass to the executable
     args: Vec<CString>,
-}
-
-/// Type for (automatic) [Executable] factories
-///
-/// We need to be able to generate [Executable]s in different ways. We might
-/// want to create one from command line arguments, or we might create one from
-/// an iterable, or we might use it for stubbing.
-///
-/// As such, we have various functions that create [Executable]s. We term these
-/// "Executable Factories." We define one of these to be "Automatic" if it takes
-/// no arguments.
-///
-/// The [main](crate::main) function can use [AutoExecutableFactory]s during
-/// runtime to create [Executable]s.
-pub type AutoExecutableFactory = fn() -> ExecutableFactoryResult;
-
-/// Convinience type for the result of an [Executable] factory
-///
-/// Creating an [Executable] may succeed or may fail. A [Result] is thus
-/// returned with the status. For convinience, this type aliases to the result.
-pub type ExecutableFactoryResult = Result<Executable, ExecutableFactoryError>;
-
-/// Error for [Executable] factories
-///
-/// When creating an [Executable], functions might run into errors with finding
-/// the parameters needed. This `enum` supplies error codes for the different
-/// possibilities.
-///
-/// Notice that a path cannot be malformed - it's just a string. Moreover, no
-/// arguments being found isn't an error - it just means the process will be
-/// created with no arguments. As such, those possibilities are not present
-/// here.
-#[derive(Debug)]
-pub enum ExecutableFactoryError {
-    /// Path could not be located
-    PathNotFound,
-
-    /// Comamnd line argument couldn't be parsed. The `position` is
-    /// the zero-indexed number of the command line argument that failed, and
-    /// `content` is the content of the string.
-    ArgMalformed { position: usize, content: String },
 }
