+++ src/permission/mod.rs
@@ -8,12 +8,7 @@
 //! them from various sources. It's tied to the Unix permission model, where
 //! each user has a user id, a group id, and a list of secondary group ids.
 
-pub mod commandline;
-pub mod environment;
-pub mod iterator;
-pub use commandline::from_commandline;
-pub use environment::from_environment;
-pub use iterator::from_iterator;
+pub mod factory;
 
 use libc::{gid_t, uid_t};
 use std::collections::HashSet;
@@ -32,44 +27,4 @@ pub struct Permission {
     /// A set of secondary group ids, which may or may not contain the primary
     /// group id itself
     secondary_gids: HashSet<gid_t>,
-}
-
-/// Type for (automatic) [Permission] factories
-///
-/// We need to be able to generate [Permission]s in different ways. We might
-/// want to create one from command line arguments, or we might create one from
-/// the environment the process is running in.
-///
-/// As such, we have various functions that create [Permission]s. We term these
-/// "Permission Factories." We define one of these to be "Automatic" if it takes
-/// no arguments.
-///
-/// The [main](crate::main) function can use [AutoPermissionFactory]s during
-/// runtime to create [Permission]s.
-#[allow(dead_code)]
-pub type AutoPermissionFactory = fn() -> PermissionFactoryResult;
-
-/// Convinience type for the result of a [Permission] factory
-///
-/// Creating a [Permission] may succeed or may fail. A [Result] is thus returned
-/// with the status. For convinience, this type aliases to the result.
-pub type PermissionFactoryResult = Result<Permission, PermissionFactoryError>;
-
-#[derive(Debug)]
-#[allow(dead_code)]
-pub enum PermissionFactoryError {
-    /// UID could not be located
-    UIDNotFound,
-    /// Primary GID could not be located
-    PrimaryGIDNotFound,
-    /// Secondary GID list could not be found
-    SecondaryGIDNotFound,
-
-    /// Parse error for UID, where `content` is the string we tried to parse
-    UIDMalformed { content: String },
-    /// Parse error for Primary GID, where `content` is the string we tried to
-    /// parse
-    PrimaryGIDMalformed { content: String },
-    /// Parse error for a Secondary GID, where `content` is the failing string
-    SecondaryGIDMalformed { content: String },
 }
